Update and rename Hardhat.yml to ci-web3-gamefi.yml (#55)
Rename the CircleCI configuration file for the web3 GameFi project and update the referenced CircleCI project URL in the Docker auth comment.

CI:

Rename the CircleCI config from Hardhat.yml to ci-web3-gamefi.yml to reflect the web3 GameFi project.
Update the CircleCI settings URL in the config comment to point to the correct GitHub project for environment variables.
Signed-off-by: Dargon789 <64915515+Dargon789@users.noreply.github.com>